cleanup and separate RequestExecutor.swift
